requirements.txt
gensim>=4.0.1
numpy>=1.20.3
scipy>=1.16.5
python-socketio>=5.2.1
sklearn==0.0